lldpad: bonding support for 802.1Qbg
Added the bonded interfaces to the list of ports. Right now bond masters
are chained by bond_porthead variable. This prohibits creating an agent
on this bond master interface.

With this patch the bond master is also added to the list of interfaces.
The special bond master archor variable bond_porthead is not used
anymore. Is is removed (together with dependend functions) with a
later patch.

Now use the bonding interface and the slave interfaces at the same
time. Set configuration file correctly. Interface bondX allows EVB
and VDP protocols when configured correctly. In this case the slave
interfaces must not run both protocols.

Correct configuration setting is prereq for correct operation.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>