lldpad: DCBX: set enable and willing bits consistently on all subtypes
Set APP data enable and willing bits for all subtypes to true at
init time. Before this only the FCoE APP subtype was configured
this way creating a confusing view of the real enable/willing bits
in the APP data.

With this change 'dcbtool' output matches what is seen on the wire.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>