lldpad: 8021qaz: removing a TLV does not update MIB
When the peer removes a TLV (ETS, PFC, or APP) the 802.1Qaz
module does not update the local MIB. This results in the
old peer data persisting even though it is no longer in the
received pdu.

This resolves the issue by clearing the local MIB even in
the case of a NULL ptr indicating no mib received.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>