ecp tlv build patch
This patch adds a function to append a TLV at the end of the
transmit buffer when the ecp data units are constructed. When the
transmit buffer is full, the data is transmitted instead of being
dropped. Also removes a memory leak when a packed tlv can not
be added.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>