ecp use static transmit buffer
This patch replaces the dynamic allocation of the transmit buffer
by a static allocation. The transmit buffer is created each time
the ecp payload is constructed. It is deleted after the frame has
been sent.
Now a static send buffer is used, avoiding the constant malloc
and free and its error handling.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>