ecp use static receive buffer
This patch replaces the dynamic allocation of the receive buffer
by a static allocation. The receive buffer is created each time
an ecp payload is received. It is deleted after the frame has
been processed and handed over to the upper layer protocol.
Now a static receive buffer is used, avoiding the constant malloc
and free and its error handling.

Also remove the badFrame state variable. It was set, but never
used.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>