vdp null pointer on vsi exit fix
RHEL Beta 6.4 migration stress testing revealed an ugly
race condition leading to an lldpad core dump on VSI
exit processing.

A command to terminate an association is received and
send to the switch. During switch processing another
command from the application is received to associate
again. The VSI profile local change timer is started.
When the switch confirmation for the disassociate is
received, the profile is removed from the list.
However the local change timer was not cancelled
and when it fired the pointer to the profile
might be valid.

Add code to cancel the local change timer before
the vsi profile is removed.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>