vdptest connect to lldpad broken
With the removal of the /var/run/lldpad.pid file, vdptest
connects to lldpad via the liblldp_clif.so library.
It sends a ping command to llpdad and expects a pongXXXXX
reply back, with XXXXX being lldpad process identifer.
The parsing of the reply fails, it is prepended by some
characters. Also the reply string is not null terminated
and might have trailing characters.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>