liblldp_clif add command to return PID of lldpad
When lldpad is invoked with command line option -p, no
/var/run/lldpad.pid file is created. External tools which
need to know the pid of lldpad, need to connect to the
control interface of lldpad and send a ping command
to retrieve the pid.

This patch adds command line option -p to the lldptool to
display the pid of lldpad. The leading PPONG string is
stripped.
Example:
24112

Same behavior for interactive mode. Each ping command
returns the PID of lldpad.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>