ecp fix unintended retransmit of ecp messages
The rework of the ecp protocol introduced an unintended
retransmit of all ecp messages even when the ecp message
was acknowledged. The ecp message sequence number was not
stored in the vsi-profile data and remained zero.
vsi-profile data has been resend and acknowledged many times.

This caused no functional issue, but an increased traffic
between llpdad and switch.

With this patch the previous behavior is restored.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>