vdp add protocol switch
Add a protocol switch for the vdp protocol version being used
on an interface. The netlink message received from a EVB user
such as libvirtd will be directed to the vdp module handling
the protocol version running on a particular interface.

The vdp protocol versions supported are draft 0.2 and draft
2.2 (ratified standard). The switch facing the interface
determines the protocol version being, lldpad simple reacts
using the proper protocol version.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>