lldpad: Document shutdown behavior
Document the fact that lldpad does not undo any of the
configuration that it has performed when it terminates.
The main use cases for termination are restarting due
to an update and during system shutdown. In both cases,
things are probably best left in their existing state.
This patch documents this behavior.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Jack Morgan <jack.morgan@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>