lldpad: 8021qaz: Use DCB_GDCBX cmd to check for IEEE support
Today in lldpad the IEEE 802.1Qaz module checks for DCB support in the
driver via a DCB_CMD_GCAP netlink msg.

But this message is really built for the older CEE version of DCBX. In
./include/net/dcbnl.h it is listed in the CEE ops section and it uses
the CEE defines such as *_PG_* instead of including the more precise
*_ETS_* attributes. It also does not give any indication of the
firmware status.

A better check for IEEE support implemented here uses the
DCB_CMD_GDCBX command which has an explicit define for IEEE/CEE and fw
or host managed.

    #define DCB_CAP_DCBX_HOST               0x01
    #define DCB_CAP_DCBX_LLD_MANAGED        0x02
    #define DCB_CAP_DCBX_VER_CEE            0x04
    #define DCB_CAP_DCBX_VER_IEEE           0x08
    #define DCB_CAP_DCBX_STATIC             0x10

With this check we can do an explicit check for IEEE and devices that
would like to be managed by the host.

However to work around some drivers which do not set the HOST bit even
though they can be managed by the host we check for the lack of the
MANAGED bit. This works on all drivers across all kernels AFAIK.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>