lldptool add support for VDP TLV
Add support for the command 'lldptool -i ethx -t -gncb -V vdp'.
This patchs redirects the collection of the data and the
pretty print of the collected data to the vdp module.
This is a hack and a quick workaround. It overwrites
the lldp_mand module id by the lldp_vdp module id.

In a perfect world a vdptool command would exist to do this.
I will keep this todo on my backlog.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>