lldpad: Primarily identify netdevs by ifindex
Presently, lldpad principally uses ifname to identify netdevs.
This can result in problems when netdevs are renamed by udev.
This problem is most easily reproduced by frequent driver
loading and unloading. There are races that happen and
sometimes lldpad will capture the netdev name before udev
renames it. Since the ifindex does not change for the life
of an interface, it provides a more positive identification,
so use it as the primary identifier. This will result in
more calls to convert ifname to ifindex, but that overhead
should be acceptable.

Even with this patch, ifname is still used for some things
and should probably be reconsidered for everything except
configurations, which need to use a persistent identifier.

V2: Modified to account for upstream changes

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Jack Morgan <jack.morgan@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>