lldpad: Fix mand_data list corruption
In an error path, a mand_data structure is freed
while still on a list. Remove the structure from
the list before freeing it. Also call mand_free_tlv
because just because mand_bld_tlv returns an error
doesn't mean that it didn't allocate some tlvs
that need to be freed.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>