vdp22 add more support for vdp22 protocol
This patch adds more support to create a VDP22 protocol VSI proile and
to query the result of a vdp22_request command from the buddy.
A vdp22_status function is introduced to collect the necessary
vsi profile data when a GETLINK message is received by the buddy.
As in vdp22_query, the interface selects the protocol version being
run on this interface.
All VSI profile commands will be handled by the vdp22_addreq()
function, it is the entry point to the VSI profile state machine
code.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>