vdp22 add timeout calcuation
The evb22 modules sends the ECP retry and ECP RTE values to the
vdp22 module. The vdp22 module calculates the timeouts for
- waitdelay
- response waitdelay
- keep alive
in microseconds. The response waitdelay is the value added to
the waitdelay when a station arms the transmit time out value.
For details see IEEE 802.1Qbg ratified standard 41.5.5.7 and
41.5.5.9.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>