lldpad: do not require active TLVs to configure attributes
No longer require TLV data structures to be initialized to read/write
configuration. This allows managing ports before link is established.

In this patch we update the command handlers to read the config file
on gets which works because sets keep the config file in sync. The set
operations are updated by removing the requirement of existing data
structures. If the data structure is not available yet the config value
is written to the configuration file only.

Because the sets may occur before ifup() any checks for valid ports
must be done in the set handler path. For 8021qaz this means verifying
the port supports DCB.

The VDP mode attribute is AFAIK the only attribute that can not
be set or queried before ifup(). This is because the mode values are
not stored in the config file. This is left for future work if the
VDP mode values need to be configured prior to ifup(). At this time
I'm not sure this would be useful.

One other change is attributes without values in the config file
will be left blank when displayed with the '-c' option. I believe this
is more in spirit with the intent of the option. One example here
given on a tap device that has not been brought online yet and has
no capabilities configuration value.

# ./lldptool -t -g ncb -i tap0 -V EVBcfg -c
fmode=bridge
capabilities=
vsis=3295
rte=15
enableTx=no

I tested many of the 8021qaz, dcbtool, VDP, and EVB commands along
with some of the other TLV types.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>