lldpad: correct IEEE DCBX capabilities check
Only need to check if the agent is managed by firmware otherwise
the software lldpad agent can drive the IEEE DCBX version.

Also if the hardware does not support the IEEE DCBNL operations
return an error operation not supported and abort initializing
the IEEE DCBX module. The theory here is if the hardware can't
support this query it likely won't support the other IEEE DCBNL
operations.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>