lldpad: check for existence of device before allowing set
With recent change to allow configuration of devices when the
interface is down the check to verify the device exists was
broken. This patch fixes it.

commit 8a44162fab3ea4bb7e5d23077766f7cfd4bcc440
Author: John Fastabend <john.r.fastabend@intel.com>
Date:   Mon Jul 8 12:12:08 2013 -0700

    lldpad: do not require active TLVs to configure attributes

Fix this by checking that first the ifindex can be looked up
and second that it is one of the valid lldp device types.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>