vdp22 protocol support for station state machine
This patch adds VDP protocol support for the IEEE 802.1 Qbg
ratified standard. It implements the station state machine
and uses the ECP protocol for communication to the switch.
The destination mac address is the nearest customer bridge.

Both versions of the VDP protocol are supported, level 0.2
(predraft) and ratified standard (also referred to as vdp22).
Which version of the protocol is used is determined by the
configuration file.

A later patch will add the vdp22 protocol bridge state machine.
With vdp22 protocol support for the bridge side, tests can
be run on a special dual network namespace setup.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>