vdp22 add bridge resource allocation code
VDP22 bridge simulation code. For now return true all the time.
When configured with --enable-debug option
sppecial combination of input parameters trigger errors decribed
in the code comment.

Will be replaced by a lldpad configuration file section to allow
rejection and acception of VSI profiles on a configurable bases.
For now this is good enough for basic state machine testing.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>