vdp22 protocol support for bridge state machine
This patch adds VDP protocol support for the IEEE 802.1 Qbg
ratified standard. It implements the bridge state machine
and uses the ECP protocol for communication to the station.
The destination mac address is the nearest customer bridge.

With vdp22 protocol support for the bridge side, tests can
be run on a special dual network namespace setup.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>