lldpad: Fixup pid file handling
Using 'flock' on the PID file is completely pointless, as the
fd will be closed afterwards. And we're binding to an abstract
socket for CLI commands prior to that, which already takes care
of having the daemon only run once.

So remove the 'flock' code and ensure that the PID file is
correctly removed upon exit.

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>