lldp: Do not print 'l2_packet_receive - recvfrom' when network is down
If the network interface is down it's perfectly valid to get a
ENETDOWN error during recvfrom. So that's nothing to worry about
and the error message should be suppressed here.
Plus it doesn't actually carry enough information to figure out
_which_ interface is down.

Those who care can always increase the logging level to get to
the desired information.

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>