liblldp_clif function clif_getpid fix wrong buffer length
The function clif_getpid() returns the PID of a running lldpad process.
It uses the function clif_request(). However the buffer length value
to indicate the size of the receiving buffer was not set.
This resulted in a random and wrong buffer size and sometimes
caused the function to fail. The returned string did not
fit into the receiving buffer (which was large enough) due to a
buffer size too small to store the reply.
This patch fixes the issue by assigning the value len the correct
value of the receive buffer size.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>