vdp set size of filter data list to zero
The VDP protocol draft 0.2 does not support the
return of changed VLAN identifiers from the switch to
the station. There is no need to return changes in
the filter data to the originator of the VSI request.
So set the size of the filter information data to
zero and nothing is returned.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>