vdp22 support tracing for manager id
With VPD22 protocol the manager identifer can be up
to 16 bytes containing non-printable characters. To trace
a VSI protofile and display a printable manager identifer
and also maintain backward compatability with draft 0.2
(where the manager id was a 1 byte number) and new function
is introduced.
It prints 1 byte manager ids as a number and other manager
as text when they consist entirely of printable characters.
If not the manager id is printed as a UUID.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>