vdp22 remove pcp and ps from structure definition
Remove the Priority code point (PCP) and PCP Significant (PS)
members from the structure fid22. Both members are now stored
in the upper bits in the vlan identifier. This is the same format
as in the TLV definition.

Compare MAC address and group identifiers in the filter
data only when the filter information format indicates
valid data.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>