vdp22-bridge support for error generation
For testing purposes the lldpad process can run a bridge.
In normal environment it returns always ok. When compiled
with BUILD_DEBUG the bridge processing code uses certain
type identifier and type version identifiers to provoke
error and return these error to test error handling on the
station side.
The scheme is enhanced to provoke more errors for testing.

This is only temporary until the lldpad configuration file
is enhanced by a bridge configuration section to describe
which VSI profile to accept and deny.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>