vdpnl: initialize vdpnl_mac to prevent segfault
In the vdptest man pages the following profile is given as an example:

	name=thomas2,mgrid=1,typeid=123452,typeidver=1,hints=none\
		uuid=a1412857-60f7-4ce1-e95a-2164943f53dd,\
		map=2-52:54:00:8e:50:53-9999,map=0/10-52:54:00:8e:50:54-8888

When this profile is used as a basic test, it causes lldpad to segfault
and vdptest to fail. The command used is:

	sudo ./vdptest -i eth5 -Aname=thomas2,mgrid=1,typeid=123452,\
	typeidver=1,hints=none,uuid=a1412857-60f7-4ce1-e95a-2164943f53dd,\
	map=2-52:54:00:8e:50:53-9999,map=0/10-52:54:00:8e:50:54-8888\
	-a thomas2,w=10,r=2

If vlan == 0 then vdp22_alloc_vsi() sets the VDP22_RETURN_VID flag and
in vdp22_status the following code is performed (and maclist == NULL):

	if (p->flags & VDP22_RETURN_VID) {
		vsi->maclist->vlan = p->fdata[0].vlan;
		...
	}

The attached changes assume that there is only one MAC/VLAN pair but it
seems consistent with the existing code.

Signed-off-by: Philbert Lin <philbert.lin@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>