lldpad: do not set rx state from TX initialization
When the TX state is initialized it was incorrectly putting the RX
state in a non-operational state. This resulted in the TLVs being
cleared.

If the admin status was RX only then each time the select loop fires
the TX state was initialized, clearing the RX state and free'ing any
queued up received PDUs. With this fix PDU queries work in RX-only
mode. This patch resolves the issue, the following should work
correctly:

# lldptool -L -i eth0 adminstatus=rx
# lldptool -t -i ethx -n
Chassis ID TLV
        MAC: 00:22:44:55:66:77
Port ID TLV
        MAC: 00:22:44:55:66:77
Time to Live TLV
        120
End of LLDPDU TLV

A follow up patch should abort the multiple initializations occuring
when tx mode is disabled.

Reported-by: Owen Dunn <osd1000@cam.ac.uk>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>