vdp22 support variable number of VLAN/QoS returns
Protocol vdp22 allows the switch to return vlan identifier and
quality of service (QoS) identifiers to lldpad. Lldpad will receive
these values (if specified in the filter information section) and
return them to the originator.
Right now only one VLAN/QoS pair is supported. This patch allows
for a variable number of VLAN/QoS reception. Allocate enough
memory when the filter information is received from the switch,
copy the information and forward it to the originator and finally
delete the memory. This is done once at VSI association.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>