vdp22 add support for command line interface
Add support to send VSI commands from an interface which
attached to lldpad and uses the control interface to
send data back and forth to lldpad.
The functions vdp22_request() and vdp22_status() get an
additional parameter to identifier the command interface
used. A value of 0 stands for the old netlink interface
messages used for draft 0.2 and a value of 1 stands
for the control interface attached to by a client tool
such lldp_tool.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>