vdp22 call back to client support
Support response from lldpad to client program attached
to the control interface of lldpad. The response is waited for
by the client program to get the response of the issued VSI
association commands.
This path is also used when the switch issues an de-association
command to terminate a VSI association.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>