clif support for sending vdp22 commands to lldpad
Add support to send IEEE 802.1 Qbg ratified standard VDP protocol
commands to the VDP22 module.

3 new functions are introduced and included in the liblldp_clif library:
clif_vsi: Sends an VSI association request to lldpad.
clif_vsievt: Wait for an event from lldpad for a certain time. The event
	message contains the switch response for the VSI request.
clif_vsiwait: Combines clif_vsi and clif_vsievt functions.

These functions operate on the VDP22 module and can not be used to
address the VDP module which implements the IEEE 801.2 Qbg draft 0.2.
The interface for draft 0.2 is the netlink message interface.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>