vdp handle dis-assoc comand without profile
RHEL 7.0 stress testing releaved a race condition
during migration. The VM is migrated away and
lldpad gets notified by the switch as it receives a
dis-assoc command for that VM's VSI. lldpad sends a notification
to libvirtd and removes the profile. Libvirtd sends another
dis-assoc command for this VM which lldpad forwards to the
switch. The switch responds with an ACK. However this takes
some time and libvirtd does not always wait long enough for the
result. In these case Libvirtd assumes its dis-assoc failed and
sends another assoc command to reverse the migration.

The dis-assoc command received from libvirtd (why do we get
it in the first place, lldpad just notified libvirtd) is valid
even if it does not makes sense. Lldpad knows the profile is been
removed. There is no need to send the dis-assoc command to the
switch. So we simple add the DIS-ASSOC request with successful
state to the list of VSIs and return the result when libvirtd
queries for success.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>