clif and vdp22_clif_cmd send correct tlvid
This patch fixes the invalid tlvid sent from the liblldp_clif
library function clif_vsi() and clif_vsiwait() to the
vdp22 module. Until now the tlvid was the module id which is
wrong. Now the tlvid is the VSI association command which
is correct.

The library interface functions are extended to specify
the tlvid.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>