VDP: vdptool first version
This is the first version of a vdp command line interface
tool to send and retrieve data to the vdp22 module.
This tool follows similar concept as the lldptool.
The command line options are similar and some intended
functionality (such as -n to retrieve neighbor inforamtion,
that is tlv data send by bridges) is not yet implemented.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>