lldp: automake fixes for dist/distcheck
This fixes the dist and distcheck automake targets for making release
archives.

All missing header files that have been added to git are listed in
noinst_HEADERS, so they get added to the release archive but are not
installed.  Some of these ( qgb_vdp22_(cmds|clif).h ) may be part of the
client API and need to be moved to lldpad_include_HEADERS?

The srcdir prefix was removed from references to liblldp_clif.la, it's
not needed and breaks builds where the output dir is different from the
source dir as this is a generated file.

The liblldp_clid-vdp22.3 man page was added.

The qbg22sim and vdptest man pages were added to the release archive,
but still should only be installed if debug is configured.

lldpad.init was added to dist_noinst_DATA to add to the release

The custom systemd unit file install rule was replaced with a dist_DATA
definition, letting automake build the rules to handle them.  Before
they were being left out of the release archive.

Same thing with the bash completion files, just tell automake where they
go.

Compared to git-archive, make dist now gets almost everything.
I left the test data files out for now, even though the binaries and
man pages are being included.

Signed-off-by: Chris Leech <cleech@redhat.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>