lldpad: switch from sysv to posix shared memory apis
The use of SysV shared memory, to pass state between running instances of
lldpad in the initramfs and then from the root fs, is difficult to work
with from a security policy.  When lldpad runs in the initramfs there is
no security policy loaded.  Then when it's restarted after an SELinux
policy has been loaded, there is no way to correct the context on the
already existing shared memory segment.  This would result in the need
for an overly permissive policy for lldpad.

By switching to POSIX APIs the segment is mapped from a tmpfs file with
a directory entry under /dev/shm/.  This lets us add a file contents
entry to the SELinux policy that matches that path, and a proper
security context can be restored to it before restarting lldpad.

Signed-off-by: Chris Leech <cleech@redhat.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>