memleak on received TLVs from modules
Most of the TLV modules that have an rchange handler for received TLVs
seem to get the return values wrong, returning 0 or TLV_OK without
freeing or storing the unpacked TLV to be freed later. That leaks the
allocation, as rxProcessFrame believes the module has claimed ownership.

In a test setup, it's probably easiest to see by enabling some TLV type
on one side of a connection only.  Or, any unexpected TLV that doesn't
get handled will be erroneously leaked by the EVB modules.