automake: introduce make check
Automake supports using 'make check' to execute so-called unit tests.  These
are used to do basic sanity checking of the code (prove that it works, etc).

This commit introduces an initial user which checks the OID string printed
by the mgmt-addr print routine.

The test will fail at this point (because there is a bug in the OID print
routine).

Signed-off-by: Aaron Conole <aconole@redhat.com>