ecp22: deinit l2_packet_data before freeing ecp22
When a link is shut, an ecp22 structure is removed for the
interface. Ecp contains an l2_packet_data which opens a socket
that needs to be closed via l2_packet_deinit before the ecp22
structure can be freed.

Failure to close the socket causes a segfault once lldpad has
exceeded the max open file descriptors.