autobuild: remove hardcoded flags, test instead
The original project was targeted only at building via 'gcc', but
it is possible that other compilers are used.  Additionally, a
distro builder might not want a default of error messages to pop
out of a build.

To support this, change the default flags away so that a user may
switch to a compiler that doesn't support the warning flags.

Additionally, the error conversion should be an option - changes
in compilers might cause warnings to show in code that previously
had none.  Ideally, this would be fixed immediately.  In practice,
it is a burden on distro maintainers to be constantly dealing
with these errors.  This now is a flag that gets tested by the
compiler.

Cc: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Cc: Tomasz Kloczko <kloczko.tomasz@gmail.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>