tx: when operating in rx-only mode don't send a port shutdown pdu
Currently, lldpad correctly transmits a shutdown pdu with ttl = 0s when
transitioning from rxtx or txonly to rxonly.  However, when we shutdown
lldpad it will transmit a shutdown pdu even if the port is configured
to rxonly mode.  For some implementations of LLDP this can create
a confusing state and lead to issues in the network.  Correct this by
only transmitting a shutdown PDU when going from any transmit mode to
a receive only mode, and don't transmit PDUs on shutdown if the port
agent isn't configured to transmit.

Reported-by: Matthew Whitehead <mwhitehe@redhat.com>
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1905210
Signed-off-by: Aaron Conole <aconole@redhat.com>