lldpad: ethtool netlink accessors for MAC merge layer
In the Linux kernel, the state associated with what's communicated via
the Additional Ethernet Capabilities TLV is held in the interface's
MAC Merge (MM) layer. This is accessed through the ETHTOOL_MSG_MM_GET
and ETHTOOL_MSG_MM_SET netlink messages. We need the GET to query the
kernel state in order to figure out what to advertise, and the SET to
reconfigure ourselves after we've figured out what the link partner has
advertised.

The lldpad program currently links with libnl-3.0, but not with the
generic netlink extension of the library (libnl-genl-3.0) which is
required for talking to the ethtool netlink interface.

Create a getter function and 3 setters in lldp_ethtool.{c,h} which
correspond with the API that will be needed by future changes.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>