lldpad: ieee8023: add Additional Ethernet Capabilities TLV logic
IEEE 802.3-2018 clause 79.3.7 Additional Ethernet Capabilities TLV
defines a 2 octet payload with 4 fields:

- preemption capability support
- preemption capability status
- preemption capability active
- additional fragment size

The additional fragment size is the simplest - what the link partner
advertises is what the local device must use.

The purpose of this TLV is to enable TX preemption (i.o.w., to allow
preemptible packets to be sent). Each link partner decides when to
enable TX preemption for themself, based on what was advertised.
We do not control what the link partner does/enables.

This kind of handshake is necessary because preemptible packets use a
custom set of Start-of-Frame-Delimiter values (generically called SMD -
Start-of-mPacket-Delimiter). Link partners which lack a MAC Merge layer
will just see these packets as error frames.

My intention is to have a plug-and-play system which ends up enabling
the maximum set of capabilities, with no configuration. In fact, to the
best of my knowledge, this is kind of the only thing that's possible to
achieve anyway.

When we support the MAC Merge layer and the link partner supports it
too (confusingly called "preemption capability support" in IEEE 802.3),
we go ahead and try to enable preemption in the TX direction. The link
partner may do that too or may do something else.

The preemption verification is a feature through which the port, once TX
preemption was enabled, will generate some special SMD-V frames and wait
for SMD-R replies from the link partner. When an SMD-R frame comes back,
the verification state machine comples, and TX preemption is not only
enabled, but also active.

Using the verification feature allows us to ensure that, even though we
enabled TX preemption without explicit agreement from the LP, we have a
backpressure mechanism which prevents us from sending packets that the
LP is simply going to drop as invalid.

To support a link partner which does something similar to us, we need to
enable the RX side of the pMAC (the entity which processes preemptible
packets and SMD-V/SMD-R verification frames) as soon as we see that the
LP has advertised support for preemption. The LP might be transmitting
SMD-V frames already (it saw our TLV first). Luckily, IEEE defines the
verifyLimit constant of 3 SMD-V retries until a SMD-R is received.
The interval between retries is verifyTime (ranges from 1 to 128 ms).

I chose to max out the verifyTime on the local device, to give the link
partner the maximum possible amount of time - 384 ms - to respond.

To recap:

- When the "preemption capability support" bits from both our the LP
  device are true, we enable TX preemption + verification

- Doing that will set the "preemption capability status" bit in our
  subsequent TLVs. This is purely informational for the LP from our POV
  (we don't care what it does with it).

- If the verification succeeds (LP must respond outside of LLDP
  protocol), the "preemption capability active" bit will be set,
  otherwise it won't.

- It is perfectly valid to have TX preemption enabled/active in one
  direction but not the other

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>